+++ src/main.rs
@@ -62,7 +62,7 @@ fn main() -> Result<()> {
     let (dispatcher_tx, dispatcher_rx) = crossbeam_channel::unbounded();
     dispatcher_tx.send(Composer::new("my-iot::start").type_(MessageType::ReadNonLogged).into())?;
     let mut all_txs = vec![core::persistence::thread::spawn(db.clone(), &dispatcher_tx)?];
-    all_txs.extend(services::spawn_all(&settings, &db, &dispatcher_tx)?);
+    all_txs.extend(core::services::spawn_all(&settings, &db, &dispatcher_tx)?);
     core::dispatcher::spawn(dispatcher_rx, dispatcher_tx, all_txs)?;
 
     info!("Starting web server on port {}…", settings.http_port);